feat: add `leaderboard` and `attempt` endpoint
`Leaderboard` is an endpoint with purpose to display best attempts on
certain quiz. For now, the ranking is only based on attempt so user with
multiple attempt may appear multiple times. The user-based ranking would
be implemented soon.

`Attempt` is and endpoin which serve attempts done by user. This could
also list attempts done based on `quiz_id` or `user_id`