feat: complete `leaderboard` endpoint handler
This allows user to get the list of best attempt that has been done in a
particular quizzes. Leaderboard is determined by user's score on an
attempt. If there're attempts with equal score, the one with smaller id
would be prioritized (since it implies the time of that attempt being
recorded)