fix: wrong sorting result
This somehow caused by a condition that always yields true, making the
algorithm always choose the first position as the position that needs to be
changed.

I'm pretty sure I used the one that did the sorting right before
committing it. I even didn't changed anything here before! Am I drunk or
something when I was making the previous commit to this file?